# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "http server api docs",
        "title": "admin api docs",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/menu": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "list of menus by paging and conditions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "list of menus by query parameters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "endTime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "keywords",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "parentId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "startTime",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ListMenusReply"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "submit information to create menu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "create menu",
                "parameters": [
                    {
                        "description": "menu information",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateMenuRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.CreateMenuReply"
                        }
                    }
                }
            }
        },
        "/api/v1/menu/options": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get role options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "get role options",
                "parameters": [
                    {
                        "type": "boolean",
                        "name": "onlyParent",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.OptionsReply"
                        }
                    }
                }
            }
        },
        "/api/v1/menu/routes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "list routes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "list of routes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Result"
                        }
                    }
                }
            }
        },
        "/api/v1/menu/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get menu detail by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "get menu detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.GetMenuByIDReply"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update menu information by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "update menu",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "menu information",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateMenuByIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UpdateMenuByIDReply"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete menu by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "delete menu",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.DeleteMenuByIDReply"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.CreateMenuReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "id": {
                            "description": "id",
                            "type": "integer"
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
        "types.CreateMenuRequest": {
            "type": "object",
            "properties": {
                "alwaysShow": {
                    "description": "始终显示",
                    "type": "integer"
                },
                "component": {
                    "description": "组件路径(vue页面完整路径，省略.vue后缀)",
                    "type": "string"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "keepAlive": {
                    "description": "始终显示",
                    "type": "integer"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "params": {
                    "description": "路由参数",
                    "type": "string"
                },
                "parentId": {
                    "description": "父级",
                    "type": "integer"
                },
                "perm": {
                    "description": "权限标识",
                    "type": "string"
                },
                "redirect": {
                    "description": "跳转路径",
                    "type": "string"
                },
                "routePath": {
                    "description": "路由路径",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "type": {
                    "description": "菜单类型(CATALOG-菜单；MENU-目录；BUTTON-按钮；EXTLINK-外链)",
                    "type": "string"
                },
                "visible": {
                    "description": "显示状态",
                    "type": "integer"
                }
            }
        },
...
        "types.DeleteMenuByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data"
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.GetMenuByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.MenuObjDetail"
                        }
                    ]
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.ListMenusReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "object",
                    "properties": {
                        "list": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.MenuObjPage"
                            }
                        },
                        "total": {
                            "type": "integer"
                        }
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.MenuObjDetail": {
            "type": "object",
            "properties": {
                "alwaysShow": {
                    "description": "始终显示",
                    "type": "integer"
                },
                "component": {
                    "description": "组件路径(vue页面完整路径，省略.vue后缀)",
                    "type": "string"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "description": "convert to uint64 id",
                    "type": "integer"
                },
                "keepAlive": {
                    "description": "始终显示",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "params": {
                    "description": "路由参数",
                    "type": "string"
                },
                "parentId": {
                    "description": "父级",
                    "type": "integer"
                },
                "perm": {
                    "description": "权限标识",
                    "type": "string"
                },
                "redirect": {
                    "description": "跳转路径",
                    "type": "string"
                },
                "routeName": {
                    "description": "路由名称",
                    "type": "string"
                },
                "routePath": {
                    "description": "路由路径",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "type": {
                    "description": "菜单类型(CATALOG-菜单；MENU-目录；BUTTON-按钮；EXTLINK-外链)",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string"
                },
                "visible": {
                    "description": "显示状态",
                    "type": "integer"
                }
            }
        },
        "types.MenuObjPage": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.MenuObjPage"
                    }
                },
                "component": {
                    "description": "组件路径(vue页面完整路径，省略.vue后缀)",
                    "type": "string"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "description": "convert to uint64 id",
                    "type": "integer"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "父级",
                    "type": "integer"
                },
                "perm": {
                    "description": "权限标识",
                    "type": "string"
                },
                "redirect": {
                    "description": "跳转路径",
                    "type": "string"
                },
                "routeName": {
                    "description": "路由名称",
                    "type": "string"
                },
                "routePath": {
                    "description": "路由路径",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "type": {
                    "description": "菜单类型(CATALOG-菜单；MENU-目录；BUTTON-按钮；EXTLINK-外链)",
                    "type": "string"
                },
                "visible": {
                    "description": "显示状态",
                    "type": "integer"
                }
            }
        },
        "types.Options": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Options"
                    }
                },
                "label": {
                    "description": "标签",
                    "type": "string"
                },
                "value": {
                    "description": "值"
                }
            }
        },
        "types.OptionsReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Options"
                    }
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data"
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
...
        "types.UpdateMenuByIDReply": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "return code",
                    "type": "integer"
                },
                "data": {
                    "description": "return data"
                },
                "msg": {
                    "description": "return information description",
                    "type": "string"
                }
            }
        },
        "types.UpdateMenuByIDRequest": {
            "type": "object",
            "properties": {
                "alwaysShow": {
                    "description": "始终显示",
                    "type": "integer"
                },
                "component": {
                    "description": "组件路径(vue页面完整路径，省略.vue后缀)",
                    "type": "string"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "description": "uint64 id",
                    "type": "integer"
                },
                "keepAlive": {
                    "description": "始终显示",
                    "type": "integer"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "params": {
                    "description": "路由参数",
                    "type": "string"
                },
                "parentId": {
                    "description": "父级",
                    "type": "integer"
                },
                "perm": {
                    "description": "权限标识",
                    "type": "string"
                },
                "redirect": {
                    "description": "跳转路径",
                    "type": "string"
                },
                "routePath": {
                    "description": "路由路径",
                    "type": "string"
                },
                "sort": {
                    "description": "排序",
                    "type": "integer"
                },
                "type": {
                    "description": "菜单类型(CATALOG-菜单；MENU-目录；BUTTON-按钮；EXTLINK-外链)",
                    "type": "string"
                },
                "visible": {
                    "description": "显示状态",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type Bearer your-jwt-token to Value",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}